changed Hello --> Home and into the views folder
