+++ .eslintrc.js
@@ -1,7 +1,6 @@
 module.exports = {
   extends: 'erb',
   plugins: ['@typescript-eslint'],
-  node: true,
   rules: {
     // A temporary hack related to IDE not resolving correct package.json
     'import/no-extraneous-dependencies': 'off',
